results/rebuttal_new_baselines_llmatic_missing/run-20231111-034200_LLMatic-zero_shot_donnemartin-system-design-oop-eventplanner_0_20-runs_log/eventplanner/15/LLMatic/event_management.py
class Event:
	def __init__(self, name, date, location):
		self.name = name
		self.date = date
		self.location = location
		self.customizations = {}

	def customize_event(self, customization):
		self.customizations.update(customization)

	def update_event(self, name=None, date=None, location=None):
		if name:
			self.name = name
		if date:
			self.date = date
		if location:
			self.location = location


class Calendar:
	def __init__(self):
		self.events = {}

	def add_event(self, event):
		self.events[event.name] = event

	def view_calendar(self):
		return self.events
